Correct default value for silent parameter
Having the default value to silent parameter to "False" obligates the user to pass `silent=True` at every request, since in the method `def request(self, method, *parts, **options):`, we set default values for parameters that are not None. In this case, setdefault was called for silent with value `False` and at the last iteration with value `True`, but as key already existed (was first set as False), this had no effect.